Moved out the LogEventEntity to a separate file
